Rework the solenoid API
This now categorises the solenoids into 3 groups:
1. On/Off solenoid (Designed for Y3/Y4/Y4 shift solenoids)
2. Inrush control solenoid (Designed for TCC solenoid)
3. Constant current solenoid (Designed for SPC/MPC solenoid)

## On/Off solenoid
1. Turns on for 500ms at maximum PWM
2. Holds for the rest of the time duration at a fixed low PWM

## Inrush solenoid

Controlled via a timer task

1. Turns on for 2.5ms at maximum PWM
2. Holds for the rest of the time at fixed PWM
3. Steps 1+2 are repeated at a rate of 50Hz to create pressure buckets

## Constant current solenoid
1. PWM Setting is NOT used - Instead they have a target current to be set
2. Current trimming is done every 10ms by the internal write_pwm task

This also removes some of the abstraction on the solenoid layer
(ConstantCurrentDriver is now removed) and instead the solenoid
type directly influences the control mechanism

Vref and Temperature compensation are pre-calculated at a rate of 1000hz
by the internal updater task for all solenoids. This reduces the computation
by 6x as now each solenoid individually is not doing this calculation.